Consolidate abstract backend code.
