+++ snuba/subscriptions/consumer.py
@@ -12,7 +12,7 @@ from typing import (
     Sequence,
 )
 
-from snuba.utils.streams.consumer import Consumer, ConsumerError
+from snuba.utils.streams.abstract import Consumer, ConsumerError
 from snuba.utils.streams.types import Message, Partition, Topic
 from snuba.utils.types import Interval, InvalidRangeError
