+++ snuba/utils/streams/synchronized.py
@@ -5,7 +5,7 @@ from typing import Callable, Mapping, MutableMapping, Optional, Sequence, Set
 
 from snuba.utils.codecs import Codec
 from snuba.utils.concurrent import Synchronized, execute
-from snuba.utils.streams.consumer import Consumer, ConsumerError, EndOfPartition
+from snuba.utils.streams.abstract import Consumer, ConsumerError, EndOfPartition
 from snuba.utils.streams.kafka import KafkaPayload
 from snuba.utils.streams.types import Message, Partition, Topic, TPayload
